refactor: Moved color map into FeatureControls
